requirements.txt
fastapi
uvicorn[standard]
streamlit
web3
requests
python-dotenv
pandas
urllib3<2
